LPD-84844 Inline the XmlRpcMethodUtil
